Added code to parse spreadsheets with multiple values per column/row
For example, a small TSV document showing a movie with multiple taglines would look like this:

/type/object/name	/film/film/initial_release_date	/film/film/tagline	/film/film/directed_by
8mm	1999-02-26	Some doors should never be opened.	Joel Schumacher
		You are never prepared for the truth.
		You can't prepare for where the truth will take you.
		Be careful when you kill a monster, that you don't become one yourself.

If there's no entry in the first column of a row, then that row contains
more values for the first row above it which does.

Since there's no way to tell if a spreadsheet is formatted this way, we
look for evidence that it is, and ask the user which way is right by
displaying an example and showing the two possible interpretations.

Currently it can't render the spreadsheet back for output.  That should be
upcoming.