Improved the data model, another decent increase in rendering speed
The speed boost comes from creating the content to be rendered by
getting the clone of a template Element in the DOM, filling it in,
and only then inserting the clone into the DOM and displaying it.

A further speed boost comes from pre-rendering the next element
and storing it hidden in the DOM, meaning that we only have to
find an element by ID and mark it for display.  This also eliminates
the need for explicit prefetching of images.

Ok ok, enough of these incremental improvements and on to the column
reconciliation.